Added HPP notice input controller, fixed small issues

+++ src/main/java/database/FirebaseDatabaseConnectable.java
@@ -1,8 +1,10 @@
 package database;
 
 import com.google.auth.oauth2.GoogleCredentials;
+import com.google.cloud.storage.Bucket;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
+import com.google.firebase.cloud.StorageClient;
 import exceptions.DuplicateConnectionToDatabaseException;
 
 import java.io.FileInputStream;
@@ -12,6 +14,7 @@ import java.net.URLConnection;
 public class FirebaseDatabaseConnectable implements DatabaseConnectable{
     @Override
     public boolean setupConnection(boolean isBoardOnline, String path) throws Exception {
+        FirebaseApp firebaseApp = null;
         if(isBoardOnline) {
             throw new DuplicateConnectionToDatabaseException();
         }
@@ -21,13 +24,17 @@ public class FirebaseDatabaseConnectable implements DatabaseConnectable{
                         new FileInputStream(path);
                 FirebaseOptions options = new FirebaseOptions.Builder()
                         .setCredentials(GoogleCredentials.fromStream(serviceAccount))
-                        .setDatabaseUrl("https://ghurdauriconnect.firebaseio.com")
+                        .setDatabaseUrl("https://notice-board-60616.firebaseio.com")
+                        .setStorageBucket("notice-board-60616.appspot.com")
                         .build();
 
-                FirebaseApp.initializeApp(options);
+                firebaseApp = FirebaseApp.initializeApp(options);
             } catch (IllegalStateException e) {
-                FirebaseApp.getInstance().delete();
-                setupConnection(isBoardOnline, path);
+                e.printStackTrace();
+                if(firebaseApp!=null) {
+                    FirebaseApp.getInstance().delete();
+                }
+                //setupConnection(isBoardOnline, path);
             }
         }
         return true;
